refactor/solution: Change SafeApp.Utilities to SafeApp.Core
