+++ SafeApp.AppBindings/Abstraction/IAppBindings.Manual.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Runtime.InteropServices;
 using System.Threading.Tasks;
-using SafeApp.Utilities;
+using SafeApp.Core;
 
 // ReSharper disable once CheckNamespace
