+++ SafeApp.AppBindings/AppBindings.cs
@@ -7,7 +7,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Runtime.InteropServices;
 using System.Threading.Tasks;
-using SafeApp.Utilities;
+using SafeApp.Core;
 
 // ReSharper disable UnusedMember.Global
 // ReSharper disable InconsistentNaming
