+++ SafeApp.MockAuthBindings/AuthBindings.Manual.cs
@@ -4,7 +4,7 @@ using System;
 using System.Runtime.InteropServices;
 using System.Threading.Tasks;
 
-using SafeApp.Utilities;
+using SafeApp.Core;
 
 #if __IOS__
 using ObjCRuntime;
