+++ SafeApp/AccessContainer.cs
@@ -3,7 +3,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using JetBrains.Annotations;
 using SafeApp.AppBindings;
-using SafeApp.Utilities;
+using SafeApp.Core;
 
 // ReSharper disable ConvertToLocalFunction
 namespace SafeApp
