+++ SafeApp/MData/MDataEntries.cs
@@ -2,7 +2,7 @@
 using System.Threading.Tasks;
 using JetBrains.Annotations;
 using SafeApp.AppBindings;
-using SafeApp.Utilities;
+using SafeApp.Core;
 
 // ReSharper disable ConvertToLocalFunction
