+++ Tests/SafeApp.Tests.Android/SafeApp.Tests.Android.csproj
@@ -100,9 +100,9 @@
       <Project>{a50ad45f-a945-41af-8b22-b2dcaa266eac}</Project>
       <Name>SafeApp.MockAuthBindings</Name>
     </ProjectReference>
-    <ProjectReference Include="..\..\SafeApp.Utilities\SafeApp.Utilities.csproj">
+    <ProjectReference Include="..\..\SafeApp.Core\SafeApp.Core.csproj">
       <Project>{a0450f47-b05e-4c53-9845-7b89d708d3e3}</Project>
-      <Name>SafeApp.Utilities</Name>
+      <Name>SafeApp.Core</Name>
     </ProjectReference>
     <ProjectReference Include="..\..\SafeApp\SafeApp.csproj">
       <Project>{82fd8e34-048c-4c74-b7a6-7cefc054f87d}</Project>
